backend/pyproject.toml
[project]
name = "hpc-agent-backend"
version = "0.1.0"
requires-python = "==3.11.*"
dependencies = [
    "flask==3.1.*",
    "flask-cors==5.0.*",
    "flask-sqlalchemy==3.1.1",
    "globus-compute-sdk>=4.7",
    "python-dotenv==1.0.*",
    "requests==2.32.*",
    "jinja2==3.1.6",
    "gunicorn==23.0.*",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
